refactor: 공통 기능을 분리한다.
- domain.Participant
