+++ src/main/java/domain/Dealer.java
@@ -1,35 +1,22 @@
 package domain;
 
-public class Dealer {
+public class Dealer extends Participant {
 
-    private static final int BUST_LIMIT = 21;
     private static final int MORE_CARD_LIMIT = 16;
-
-    private final Name name = new Name("딜러");
-    private final Cards cards;
+    private static final Name DEALER_NAME = new Name("딜러");
 
     public Dealer(Cards cards) {
-        this.cards = cards;
+        super(DEALER_NAME, cards);
     }
 
-    public void pick(Card card) {
-        cards.addNewCard(card);
-    }
-
-    public int getTotalScore() {
-        return cards.calculateScore(MORE_CARD_LIMIT);
-    }
-
-    public boolean isBust() {
-        return getTotalScore() > BUST_LIMIT;
-    }
-
-    public Cards getCards() {
-        return cards;
-    }
-
+    @Override
     public boolean isMoreCardAble() {
         return getTotalScore() <= MORE_CARD_LIMIT;
+    }
+
+    @Override
+    public int getTotalScore() {
+        return super.cards.calculateScore(MORE_CARD_LIMIT);
     }
 
 }
