+++ src/test/java/domain/DealerTest.java
@@ -22,7 +22,7 @@ public class DealerTest {
     @Test
     @DisplayName("새로운 카드를 추가한다.")
     void pickNewCard() {
-        Dealer dealer = new Dealer(new Cards(cards));
+        Participant dealer = new Dealer(new Cards(cards));
 
         dealer.pick(new Card(Shape.DIAMOND, Value.NINE));
 
@@ -32,7 +32,7 @@ public class DealerTest {
     @Test
     @DisplayName("딜러의 점수가 21점을 넘기면 bust 이다.")
     void bustDealer() {
-        Dealer dealer = new Dealer(new Cards(cards));
+        Participant dealer = new Dealer(new Cards(cards));
 
         dealer.pick(new Card(Shape.HEART, Value.QUEEN));
 
@@ -42,7 +42,7 @@ public class DealerTest {
     @Test
     @DisplayName("딜러의 점수가 21점 이하이면 bust 가 아니다.")
     void notBustDealer() {
-        Dealer dealer = new Dealer(new Cards(cards));
+        Participant dealer = new Dealer(new Cards(cards));
 
         dealer.pick(new Card(Shape.HEART, Value.ACE));
 
@@ -52,7 +52,7 @@ public class DealerTest {
     @Test
     @DisplayName("딜러의 점수가 17점 이상이면 더 이상 카드를 받지 못한다.")
     void noMoreCard() {
-        Dealer dealer = new Dealer(new Cards(cards));
+        Participant dealer = new Dealer(new Cards(cards));
 
         dealer.pick(new Card(Shape.HEART, Value.FIVE));
 
@@ -62,7 +62,7 @@ public class DealerTest {
     @Test
     @DisplayName("딜러의 점수가 16점 이하이면 카드를 더 받을 수 있다.")
     void isMoreCardAble() {
-        Dealer dealer = new Dealer(new Cards(cards));
+        Participant dealer = new Dealer(new Cards(cards));
 
         dealer.pick(new Card(Shape.HEART, Value.FOUR));
